Rework `BaseWrappedHierarchy` into `TemplWrappedHierarchy`.
