+++ CHANGELOG.md
@@ -23,9 +23,10 @@ Breaking changes:
     doesn't affect how components are instantiated. For example, this doesn't
     make you change your standalone config files.
   * New base class `BaseRootComponent` to be the superclass for root components.
-  * Switched a couple classes from _base_ classes to _template_ classes:
+  * Switched a few classes from _base_ classes to _template_ classes:
     * `BaseThreadComponent` -> `TemplThreadComponent`.
     * `BaseAggregateComponent` -> `TemplAggregateComponent`.
+    * `BaseWrappedHierarchy` -> `TemplWrappedHierarchy`.
   * New exported classes to help with testing: `MockComponent` and
     `MockRootComponent`.
